Fix 64 bit warning.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@12319 a1c6a512-1295-4272-9138-f99709370657

+++ apps/menus/main_menu.c
@@ -99,7 +99,7 @@ static int reset_settings(void)
 }
 static int write_settings_file(void* param)
 {
-    return settings_save_config((int)param);
+    return settings_save_config((intptr_t)param);
 }
 
 MENUITEM_FUNCTION_WPARAM(browse_configs, ID2P(LANG_CUSTOM_CFG),
